split `cmd/knit/knitgraph/graph.go` into nodes, edges and graph
because the file is bloating
